api com fire storage

+++ app/src/main/java/com/br/matchmovies/model/modelSimilar/Result.kt
@@ -13,6 +13,6 @@ data class Result(
     val release_date: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Int,
+    val vote_average: Double,
     val vote_count: Int
 )
